+++ euphonic_sqw_models/euphonic_wrapper.py
@@ -1,7 +1,10 @@
-import numpy as np
-import os
 import warnings
-from euphonic import Crystal, ForceConstants, QpointPhononModes, DebyeWaller, ureg
+from typing import Optional, Dict, Union, Tuple
+
+import numpy as np
+
+from euphonic import (ForceConstants, QpointPhononModes, DebyeWaller,
+                      Quantity, ureg)
 from euphonic.util import mp_grid
 
 # Allow Euphonic's deprecation warnings to be raised if deprecated
@@ -11,128 +14,190 @@ warnings.filterwarnings('default', category=DeprecationWarning,
 
 class CoherentCrystal(object):
     """
-    A class to calculate neutron scattering dispersions surfaces for use in Horace.
-    It must be constructed from a Euphonic ForceConstants object.
+    A class to calculate neutron scattering dispersions surfaces for
+    use in Horace. It must be constructed from a Euphonic
+    ForceConstants object.
 
     Methods
     -------
-    w, sf = horace_disp(qh, qk, ql, intensity_scale=1.0, frequency_scale=1.0)
-        Calculates the phonon dispersion surfaces `w` and structure factor `sf` at the specified
-        (qh, qk, ql) points, with optional intensity and frequency scaling factors
+    w, sf = horace_disp(qh: np.ndarray, qk: np.ndarray, ql: np.ndarray,
+                        intensity_scale: float = 1.0,
+                        frequency_scale: float = 1.0)
+        Calculates the phonon dispersion surfaces `w` and structure
+        factor `sf` at the specified (qh, qk, ql) points, with optional
+        intensity and frequency scaling factors
 
     Attributes
     ----------
-    temperature : float Quantity (default: 0 K)
-        Temperature at which to calculate phonons (used for Bose and Debye-Waller factor calculations)
-    bose : bool (default: True)
-        Whether to include the Bose temperature factor in the calculated structure factor
+    force_constants : ForceConstants
+        The force constants
+    debye_waller_grid : (3,) float ndarray or None (default: None)
+        The Monkhorst-Pack q-point grid size to use for the
+        Debye-Waller calculation
     debye_waller : DebyeWaller object
-        This is computed and set internally by the class when the computation requests it 
-        (debye_waller_grid is set and temperature is non-zero)
-    debye_waller_grid : (3,) float ndarray (default: None)
-        The Monkhorst-Pack q-point grid size to use for the Debye-Waller calculation
+        This is computed and set internally by the class when the
+        computation requests it (debye_waller_grid is set and
+        temperature is not `None`)
+    temperature : float, float Quantity or None (default: None)
+        Temperature at which to calculate phonons (used for Bose and
+        Debye-Waller factor calculations). If None no temperature
+        dependent effects will be applied
+    bose : bool (default: True)
+        Whether to include the Bose temperature factor in the
+        calculated structure factor
     negative_e : bool (default: False)
-        Whether to calculate the phonon anihilation / neutron energy gain / negative energy spectrum
-    conversion_mat : (3, 3) float ndarray (default: identity)
-        This matrix is applied to the input (hkl) q-point values before the phonon calculation
+        Whether to calculate the phonon annihilation / neutron energy
+        gain / negative energy spectrum
+    conversion_mat : (3, 3) float ndarray or None (default: None)
+        This matrix is applied to the input (hkl) q-point values before
+        the phonon calculation
     chunk : float (default: 5000)
-        If non-zero will chunk the phonon calculation into blocks of <chunk> q-points
+        If non-zero will chunk the phonon calculation into blocks of
+        <chunk> q-points
     lim : float (default: infinity)
-        A cut-off for the calculated structure factor where values above `lim` is set equal to it.
+        A cut-off for the calculated structure factor where values
+        above `lim` are set equal to it
     scattering_lengths : dict or string (default: 'Sears1992')
-        The scattering lengths (in femtometres) to use for the phonon calculations either as 
-        a dictionary with elements as keys or a string denoting a database (internal or from file)
-    asr : {'realspace', 'reciprocal'}, optional
-    dipole : boolean, optional
-    dipole_parameter : float, optional
-    eta_scale : float, optional
-        .. deprecated:: 0.4.0
-        Deprecated since euphonic_sqw_models 0.4.0 and Euphonic 0.6.0
-    splitting : boolean, optional
-    insert_gamma : boolean, optional
-    reduce_qpts : boolean, optional
-    use_c : boolean, optional
-    n_threads : int, optional
-        These are parameters used in the `calculate_qpoint_phonon_modes` method of ForceConstants
-        Type `help(fc.calculate_qpoint_phonon_modes)` where fc is the ForceConstants object you created.
+        The scattering lengths to use for the phonon calculations
+        either as a dictionary with elements as keys and floats
+        (in fm) or Quantities as values, or a string denoting a
+        database (internal or from a file)
+    verbose : bool (default: True)
+        Whether to print information on calculation progress
+    **calc_modes_kwargs
+        Any other keyword arguments (e.g asr, n_threads) will be
+        passed to euphonic.ForceConstants.calculate_qpoint_phonon_modes.
+        See the available arguments with
+        help(euphonic.ForceConstants.calculate_qpoint_phonon_modes)
     """
-    # This a wrapper around the Euphonic ForceConstants and QpointPhononModes classes to make it easier to access from Matlab
-    # It is meant to be used with a Matlab python_wrapper class and implements a horace_sqw function for use with Horace
-
-    defaults = {'debye_waller': None,
-                'debye_waller_grid': None,
-                'temperature': 0.0 * ureg('K'),
-                'bose': True,
-                'negative_e': False,
-                'conversion_mat': None,
-                'chunk': 5000,
-                'lim': np.inf,
-                'scattering_lengths': 'Sears1992',
-                'weights': None,
-                'asr': None,
-                'dipole': True,
-                'dipole_parameter': 1.0 ,
-                'eta_scale': 1.0,
-                'splitting': True,
-                'insert_gamma': False,
-                'reduce_qpts': True,
-                'use_c': None,
-                'n_threads': None,
-                'verbose': True}
-
-    def __init__(self, force_constants, **kwargs):
-        for key, val in self.defaults.items():
-            setattr(self, key, kwargs.pop(key, self.defaults[key]))
-        if kwargs:
-            raise ValueError(
-                f'Unrecognised keyword arguments {list(kwargs.keys())}, '
-                f'accepted arguments are {list(self.defaults.keys())}')
+    def __init__(
+            self,
+            force_constants: ForceConstants,
+            debye_waller_grid: Optional[Tuple[int, int, int]] = None,
+            debye_waller: Optional[DebyeWaller] = None,
+            temperature: Union[float, Quantity, None] = None,
+            bose: bool = True,
+            negative_e: bool = False,
+            conversion_mat: Optional[np.ndarray] = None,
+            chunk: int = 5000,
+            lim: float = np.inf,
+            scattering_lengths: Union[
+                str, Dict[str, Union[Quantity, float]]] = 'Sears1992',
+            verbose: bool = True,
+            **calc_modes_kwargs) -> None:
+        """
+        Parameters
+        ----------
+        force_constants
+            The force constants
+        debye_waller_grid
+            The Monkhorst-Pack q-point grid size to use for the
+            Debye-Waller calculation. If None the Debye-Waller
+            calculation is not performed.
+        debye_waller
+            This is computed and set internally by the class when the
+            computation requests it (debye_waller_grid is set and
+            temperature is not None)
+        temperature
+            Temperature at which to calculate phonons (used for Bose and
+            Debye-Waller factor calculations). If None no
+            temperature-dependent effects will be applied
+        bose
+            Whether to include the Bose temperature factor in the
+            calculated structure factor
+        negative_e
+            Whether to calculate the phonon annihilation / neutron energy
+            gain / negative energy spectrum
+        conversion_mat
+            Shape (3, 3) float ndarray. This matrix is applied to the
+            input (hkl) q-point values before the phonon calculation
+        chunk
+            If non-zero will chunk the phonon calculation into blocks of
+            <chunk> q-points
+        lim
+            A cut-off for the calculated structure factor where values
+            above `lim` are set equal to it
+        scattering_lengths
+            The scattering lengths to use for the phonon
+            calculations either as a dictionary with elements as keys and
+            floats (in fm) or Quantities as values, or a
+            string denoting a database (internal or from a file)
+        verbose
+            Whether to print information on calculation progress
+        **calc_modes_kwargs
+            Any other keyword arguments (e.g asr, n_threads) will be
+            passed to euphonic.ForceConstants.calculate_qpoint_phonon_modes.
+            See the available arguments with
+            help(euphonic.ForceConstants.calculate_qpoint_phonon_modes)
+        """
         self.force_constants = force_constants
+        self.debye_waller_grid = debye_waller_grid
+        self.debye_waller = debye_waller
+        self.temperature = temperature
+        self.bose = bose
+        self.negative_e = negative_e
+        self.conversion_mat = conversion_mat
+        self.chunk = chunk
+        self.lim = lim
+        self.scattering_lengths = scattering_lengths
+        self.verbose = verbose
+        self.calc_modes_kwargs = calc_modes_kwargs
 
-    def _calculate_sf(self, qpts):
-        if self.temperature > 0:
-            if self.debye_waller is None and self.debye_waller_grid is not None:
+    def _calculate_sf(self, qpts: np.ndarray
+                      ) -> Tuple[np.ndarray, np.ndarray]:
+        if self.temperature is not None:
+            if (self.debye_waller is None
+                and self.debye_waller_grid is not None):
                 self._calculate_debye_waller()
         phonons = self._calculate_phonon_modes(qpts)
-        sf_obj = phonons.calculate_structure_factor(scattering_lengths=self.scattering_lengths,
-                                                    dw=self.debye_waller)
+        sf_obj = phonons.calculate_structure_factor(
+            scattering_lengths=self.scattering_lengths,
+            dw=self.debye_waller)
         w = sf_obj.frequencies.magnitude
         sf = sf_obj.structure_factors.magnitude
-        if self.temperature > 0 and self.bose:
+        if self.temperature is not None and self.bose:
             bose = sf_obj._bose_factor(self.temperature)
             sf = (1 + bose) * sf
         if self.negative_e:
             w = np.hstack((w, -w))
             neg_sf = sf_obj.structure_factors.magnitude
-            if self.temperature > 0 and self.bose:
+            if self.temperature is not None and self.bose:
                 neg_sf = bose * neg_sf
             sf = np.hstack((sf, neg_sf))
         return w, sf
-        
-    def horace_disp(self, qh, qk, ql, intensity_scale=1.0, frequency_scale=1.0, *args, **kwargs):
+
+    def horace_disp(self, qh: np.ndarray, qk: np.ndarray, ql: np.ndarray,
+                    intensity_scale: float = 1.0,
+                    frequency_scale: float = 1.0
+                    ) -> Tuple[Tuple[np.ndarray, ...],
+                               Tuple[np.ndarray, ...]]:
         """
-        Calculates the phonon dispersion surface for input qh, qk, and ql vectors for use with Horace
- 
+        Calculates the phonon dispersion surface for input qh, qk, and
+        ql vectors for use with Horace
+
         Parameters
         ----------
-        qh, qk, ql : (n_pts,) float ndarray
-            The q-points to calculate at as separate vectors
-        intensity_scale : float
+        qh
+            The q-points in H to calculate
+        qk
+            The q-points in K to calculate
+        ql
+            The q-points in L to calculate
+        intensity_scale
             The factor to multiply the intensity by
-        frequency_scale : float
+        frequency_scale
                 The factor to multiply the phonon frequencies
-                by, as DFT can often overestimate frequencies
-        args: tuple
-            Arguments passed directly to the convolution function
-        kwargs : dict
-            Keyword arguments passed directly to the convolution function
+                by, as DFT can often over/underestimate frequencies
 
         Returns
         -------
-        w : (n_modes,) tuple of (n_pts,) float ndarray
-            The phonon dispersion energies as a tuple of numpy float vectors
-        sf : (n_modes,) tuple of (n_pts,) float ndarray
-            The dynamical structure corresponding to phonon energies in w as a tuple of numpy float vectors
+        w
+            Length (n_qpts,) tuple of length (n_modes,) float ndarrays.
+            The phonon dispersion energies
+        sf
+            Length (n_qpts,) tuple of length (n_modes,) float ndarrays.
+            The dynamical structure corresponding to phonon energies in
+            w
         """
         if self.chunk > 0:
             lqh = len(qh)
@@ -140,8 +205,11 @@ class CoherentCrystal(object):
                 qi = i * self.chunk
                 qf = min((i+1) * self.chunk, lqh)
                 if self.verbose:
-                    print(f'Using Euphonic to interpolate for q-points {qi}:{qf} out of {lqh}')
-                qpts = np.vstack((np.squeeze(qh[qi:qf]), np.squeeze(qk[qi:qf]), np.squeeze(ql[qi:qf]))).T
+                    print(f'Using Euphonic to interpolate for q-points '
+                          f'{qi}:{qf} out of {lqh}')
+                qpts = np.vstack((np.squeeze(qh[qi:qf]),
+                                  np.squeeze(qk[qi:qf]),
+                                  np.squeeze(ql[qi:qf]))).T
                 if self.conversion_mat is not None:
                     qpts = np.matmul(qpts, self.conversion_mat)
                 sqw = self._calculate_sf(qpts)
@@ -152,7 +220,9 @@ class CoherentCrystal(object):
                     w = np.vstack((w, sqw[0]))
                     sf = np.vstack((sf, sqw[1]))
         else:
-            qpts = np.vstack((np.squeeze(qh), np.squeeze(qk), np.squeeze(ql))).T
+            qpts = np.vstack((np.squeeze(qh),
+                              np.squeeze(qk),
+                              np.squeeze(ql))).T
             if self.conversion_mat is not None:
                 qpts = np.matmul(qpts, self.conversion_mat)
             w, sf = self._calculate_sf(qpts)
@@ -161,50 +231,47 @@ class CoherentCrystal(object):
         if intensity_scale != 1.:
             sf *= intensity_scale
         sf = np.minimum(sf, self.lim)
-        # Splits into different dispersion surfaces (python tuple == matlab cell)
-        # But the data must be contiguous in memory so we need to do a real tranpose (.T just changes strides)
-        # So we need to convert to "fortran" format (which physically transposes data) before doing ".T"
+        # Splits into different dispersion surfaces
+        # (python tuple == matlab cell)
+        # But the data must be contiguous in memory so we need to do a
+        # real tranpose (.T just changes strides)
+        # So we need to convert to "fortran" format (which physically
+        # transposes data) before doing ".T"
         w = np.asfortranarray(w).T
         sf = np.asfortranarray(sf).T
         return tuple(w), tuple(sf)
 
-    def _calculate_phonon_modes(self, qpts):
-        if self.force_constants is None:
-            raise RuntimeError('Force constants model not set')
-        return self.force_constants.calculate_qpoint_phonon_modes(qpts,
-            weights=self.weights, asr=self.asr, dipole=self.dipole, dipole_parameter=self.dipole_parameter,
-            eta_scale=self.eta_scale, splitting=self.splitting, insert_gamma=self.insert_gamma,
-            reduce_qpts=self.reduce_qpts, use_c=self.use_c, n_threads=self.n_threads)
+    def _calculate_phonon_modes(self, qpts: np.ndarray
+                                ) -> QpointPhononModes:
+        return self.force_constants.calculate_qpoint_phonon_modes(
+            qpts, **self.calc_modes_kwargs)
 
-    def _calculate_debye_waller(self):
-        if self.temperature <= 0.0:
-            return
+    def _calculate_debye_waller(self) -> None:
         if self.debye_waller_grid is None:
-            raise RuntimeError('Q-points grid for Debye Waller calculation not set')
+            raise RuntimeError(
+                'Q-points grid for Debye Waller calculation not set')
         dw_qpts = mp_grid(self.debye_waller_grid)
         dw_phonons = self._calculate_phonon_modes(dw_qpts)
-        self.debye_waller = dw_phonons.calculate_debye_waller(self.temperature)
+        self.debye_waller = dw_phonons.calculate_debye_waller(
+            self.temperature)
 
     @property
-    def force_constants(self):
+    def force_constants(self) -> ForceConstants:
         return self._force_constants
 
     @force_constants.setter
-    def force_constants(self, val):
-        if val is None or val == 'None':  # Using string 'None' to make it easier for Matlab users
-            self._force_constants = None
+    def force_constants(self, val: ForceConstants) -> None:
+        if hasattr(val, 'calculate_qpoint_phonon_modes'):
+            self._force_constants = val
         else:
-            if hasattr(val, 'calculate_qpoint_phonon_modes'):
-                self._force_constants = val
-            else:
-                raise ValueError('Invalid force constant model')
+            raise ValueError('Invalid force constant model')
 
     @property
-    def debye_waller(self):
-        return self._debye_waller 
+    def debye_waller(self) -> Union[DebyeWaller, None]:
+        return self._debye_waller
 
     @debye_waller.setter
-    def debye_waller(self, val):
+    def debye_waller(self, val: Union[DebyeWaller, None, str]) -> None:
         if val is None or val == 'None':
             self._debye_waller = None
         else:
@@ -214,28 +281,29 @@ class CoherentCrystal(object):
                 raise ValueError('Invalid Debye-Waller object')
 
     @property
-    def debye_waller_grid(self):
+    def debye_waller_grid(self) -> Union[Tuple[int, int, int], None]:
         return self._debye_waller_grid
 
     @debye_waller_grid.setter
-    def debye_waller_grid(self, val):
+    def debye_waller_grid(self, val: Union[Tuple[int, int, int], None, str]
+                          ) -> None:
         if val is None or val == 'None':
             self._debye_waller_grid = None
         else:
-            val = np.squeeze(np.array(val))
-            if np.shape(val) == (3, ):
-                self._debye_waller_grid = [int(v) for v in val]
+            val = tuple(np.squeeze(np.array(val)))
+            if np.shape(val) == (3,):
+                self._debye_waller_grid = tuple([int(v) for v in val])
                 # Reset the Debye Waller factor if it was previously set.
                 self.debye_waller = None
             else:
                 raise ValueError('Invalid Debye-Waller grid')
 
     @property
-    def conversion_mat(self):
+    def conversion_mat(self) -> Union[np.ndarray, None]:
         return self._conversion_mat
 
     @conversion_mat.setter
-    def conversion_mat(self, val):
+    def conversion_mat(self, val: Union[np.ndarray, None, str]) -> None:
         if val is None or (isinstance(val, str) and val.startswith('None')):
             self._conversion_mat = None
         else:
@@ -246,31 +314,41 @@ class CoherentCrystal(object):
                 raise ValueError('Invalid conversion matrix')
 
     @property
-    def temperature(self):
+    def temperature(self) -> Union[Quantity, None]:
         return self._temperature
 
     @temperature.setter
-    def temperature(self, val):
-        self._temperature = val if hasattr(val, 'units') else (val * ureg('K'))
+    def temperature(self, val: Union[Quantity, float, None]) -> None:
+        if val is None or val == 'None':
+            self._temperature = None
+        else:
+            if isinstance(val, (float, int)):
+                val = val*ureg('K')
+            if hasattr(val, 'units') and val >= 0*ureg('K'):
+                self._temperature = val
+            else:
+                raise ValueError('Invalid temperature')
 
     @property
-    def scattering_lengths(self):
+    def scattering_lengths(self) -> Union[str, Dict[str, Quantity]]:
         return self._scattering_lengths
 
     @scattering_lengths.setter
-    def scattering_lengths(self, val):
+    def scattering_lengths(
+            self, val: Union[str, Dict[str, Union[float, Quantity]]]) -> None:
         if isinstance(val, str):
             self._scattering_lengths = val
         elif isinstance(val, dict):
             self._scattering_lengths = \
-                {ky: (v if hasattr(v, 'units') else v * ureg('fm')) for ky, v in val.items()}
+                {ky: (v if hasattr(v, 'units') else v * ureg('fm'))
+                     for ky, v in val.items()}
         else:
             raise ValueError('Invalid scattering lengths')
 
     @property
-    def chunk(self):
+    def chunk(self) -> int:
         return self._chunk
 
     @chunk.setter
-    def chunk(self, val):
+    def chunk(self, val: int) -> None:
         self._chunk = int(val)