+++ test/euphonic_horace_test.py
@@ -5,8 +5,10 @@ import numpy as np
 import numpy.testing as npt
 import scipy.io
 import euphonic
+from euphonic import ureg
 
 import euphonic_sqw_models
+
 def get_abspath(filename, sub_dir):
     test_dir = os.path.dirname(os.path.realpath(__file__))
     return test_dir + os.path.sep + sub_dir + os.path.sep + filename
@@ -35,6 +37,9 @@ qpts = [[0.0,  0.0,  0.0],
 qpts = np.array(qpts)
 scattering_lengths = {'La': 8.24, 'Zr': 7.16, 'O': 5.803,
                       'Si': 4.1491, 'Na': 3.63, 'Cl': 9.577}
+scattering_lengths_quantity = {'La': 8.24*ureg('fm'), 'Zr': 7.16*ureg('fm'),
+                               'O': 5.803*ureg('fm'), 'Si': 4.1491*ureg('fm'),
+                               'Na': 3.63*ureg('fm'), 'Cl': 9.577*ureg('fm')}
 iscale = 1.0
 freq_scale = 1.0
 pars = [iscale, freq_scale]
@@ -58,10 +63,19 @@ def get_test_opts():
     return opts
 
 
+def get_expected_output_dir():
+    return 'expected_output'
+
+
 def get_expected_output_filename(material_name, opts):
-    fname = f"{material_name}_T{opts['temperature']}"
+    fname = f'{material_name}'
+    if opts.get('temperature', None) is not None:
+        if hasattr(opts.get('temperature'), 'units'):
+            fname += f'_T{opts["temperature"].magnitude}'
+        else:
+            fname += f'_T{opts["temperature"]}'
     if opts.get('debye_waller_grid', None) is not None:
-        fname += "_dw" + "".join([str(v) for v in opts['debye_waller_grid']])
+        fname += '_dw' + ''.join([str(v) for v in opts['debye_waller_grid']])
     if opts.get('bose', None) is not None:
         fname += '_bose' + str(opts['bose']).lower()
     if opts.get('negative_e', None) is not None:
@@ -79,13 +93,12 @@ def get_expected_output_filename(material_name, opts):
         fname += '_lim' + strlim
     fname = fname.replace('.', 'p').replace('-', 'm')
     fname += '.mat'
-    return get_abspath(fname, 'expected_output');
+    return get_abspath(fname, get_expected_output_dir());
 
 
 def sum_degenerate_modes(w, sf):
     tol = 0.1;
     rows, cols = np.shape(w);
-    diff = np.zeros(cols-1);
     summed_sf = np.zeros(np.shape(sf));
     for i in range(rows):
         sum_at = np.where(np.diff(w[i,:]) > tol)[0];
@@ -101,7 +114,6 @@ def calculate_w_sf(material_opts, material_constructor, opt_dict,
                    sum_sf=True, hdisp_args=(), hdisp_kwargs={}):
     fc = material_constructor(**material_opts)
     opt_dict['asr'] = 'reciprocal'
-    opt_dict['scattering_lengths'] = scattering_lengths
     coherent_sqw = euphonic_sqw_models.CoherentCrystal(fc, **opt_dict)
     w, sf = coherent_sqw.horace_disp(
         qpts[:,0], qpts[:,1], qpts[:,2], *hdisp_args, **hdisp_kwargs)
@@ -120,8 +132,14 @@ def calculate_w_sf(material_opts, material_constructor, opt_dict,
 
 def get_expected_w_sf(fname, sum_sf=True):
     ref_dat = scipy.io.loadmat(fname)
-    expected_w = np.concatenate(ref_dat['expected_w'][0], axis=1)
-    expected_sf = np.concatenate(ref_dat['expected_sf'][0], axis=1)
+    # Was written from Matlab - convert first
+    if ref_dat['expected_w'].dtype == object:
+        expected_w = np.concatenate(ref_dat['expected_w'][0], axis=1)
+        expected_sf = np.concatenate(ref_dat['expected_sf'][0], axis=1)
+    else:
+    # Was written from Python - use as is
+        expected_w = ref_dat['expected_w']
+        expected_sf = ref_dat['expected_sf']
     # Set acoustic structure factors to zero
     expected_sf[:,:3] = 0
     if 'negative_e' in fname:
@@ -131,11 +149,15 @@ def get_expected_w_sf(fname, sum_sf=True):
         expected_sf = sum_degenerate_modes(expected_w, expected_sf)
     return expected_w, expected_sf
 
-@pytest.mark.parametrize("material", materials)
-@pytest.mark.parametrize("opt_dict", get_test_opts())
-@pytest.mark.parametrize("run_opts", [
-    {'use_c': False, 'n_threads': 1, 'chunk': 5, 'dipole_parameter': 0.75},
-    {'use_c': True, 'n_threads': 1, 'chunk': 0},
+
+@pytest.mark.parametrize('material', materials)
+@pytest.mark.parametrize('opt_dict', get_test_opts())
+# The following options shouldn't change the result
+@pytest.mark.parametrize('run_opts', [
+    {'use_c': False, 'n_threads': 1, 'chunk': 5, 'dipole_parameter': 0.75,
+     'scattering_lengths': scattering_lengths},
+    {'use_c': True, 'n_threads': 1, 'chunk': 0,
+     'scattering_lengths': scattering_lengths_quantity},
     {'use_c': True, 'n_threads': 2}])
 def test_euphonic_sqw_models(material, opt_dict, run_opts):
     material_name, material_constructor, material_opts = material
@@ -146,14 +168,43 @@ def test_euphonic_sqw_models(material, opt_dict, run_opts):
     opt_dict.update(run_opts)
     # Don't pass None options
     opt_dict = {k: v for k, v in opt_dict.items() if v is not None}
-
     w, sf = calculate_w_sf(material_opts, material_constructor, opt_dict)
     npt.assert_allclose(w, expected_w, rtol=1e-5, atol=1e-2)
     npt.assert_allclose(sf, expected_sf, rtol=1e-2, atol=1e-2)
 
+@pytest.mark.parametrize('material', materials)
+def test_euphonic_sqw_models_defaults(material):
+    material_name, material_constructor, material_opts = material
+    w, sf = calculate_w_sf(material_opts, material_constructor, {})
+    expected_w, expected_sf = get_expected_w_sf(
+        get_abspath(f'{material_name}_defaults.mat',
+                    get_expected_output_dir()))
+    npt.assert_allclose(w, expected_w, rtol=1e-5, atol=1e-2)
+    npt.assert_allclose(sf, expected_sf, rtol=1e-2, atol=1e-2)
 
-@pytest.mark.parametrize("material", materials)
-@pytest.mark.parametrize("opt_dict", [{
+
+@pytest.mark.parametrize('material', materials)
+@pytest.mark.parametrize('temperature', [None, 0*ureg('K')])
+@pytest.mark.parametrize('opt_dict', [{
+    'debye_waller_grid': [6, 6, 6],
+    'negative_e': True,
+    'conversion_mat': (1./2)*np.array([[-1, 1, 1],
+                                       [1, -1, 1],
+                                       [1, 1, -1]])}])
+def test_euphonic_sqw_models_temperatures(
+        material, opt_dict, temperature):
+    material_name, material_constructor, material_opts = material
+    opt_dict.update({'temperature': temperature})
+    w, sf = calculate_w_sf(material_opts, material_constructor, opt_dict)
+    expected_w, expected_sf = get_expected_w_sf(
+        get_expected_output_filename(
+            material_name, opt_dict))
+    npt.assert_allclose(w, expected_w, rtol=1e-5, atol=1e-2)
+    npt.assert_allclose(sf, expected_sf, rtol=1e-2, atol=1e-2)
+
+
+@pytest.mark.parametrize('material', materials)
+@pytest.mark.parametrize('opt_dict', [{
     'temperature': 300,
     'debye_waller_grid': [6, 6, 6],
     'negative_e': True,
@@ -163,7 +214,7 @@ def test_euphonic_sqw_models(material, opt_dict, run_opts):
 # Test a combination of keyword and positional arguments
 # Is required for use with pace-python (Toby/Multifit
 # allow positional only for fitting parameters)
-@pytest.mark.parametrize("iscale, freqscale, args, kwargs",
+@pytest.mark.parametrize('iscale, freqscale, args, kwargs',
         [(1.0, 1.3, (), {'frequency_scale': 1.3}),
          (1e-4, 1.0, (1e-4,), {}),
          (1.5e3, 0.4, (1.5e3, 0.4), {}),
@@ -188,8 +239,8 @@ def test_euphonic_sqw_models_pars(
                         rtol=1e-2, atol=1e-2*iscale)
 
 
-@pytest.mark.parametrize("opt_dict", [{
-    'temperature': 300,
+@pytest.mark.parametrize('opt_dict', [{
+    'temperature': 300*ureg('K'),
     'debye_waller_grid': [6, 6, 6],
     'negative_e': True,
     'conversion_mat': (1./2)*np.array([[-1, 1, 1],
@@ -231,8 +282,8 @@ def test_old_behaviour_single_parameter_sets_intensity_scale(opt_dict):
 
 # Units of sf have changed, test the calculated and old expected
 # values are the same apart from a scale factor
-@pytest.mark.parametrize("material", materials)
-@pytest.mark.parametrize("opt_dict", [{
+@pytest.mark.parametrize('material', materials)
+@pytest.mark.parametrize('opt_dict', [{
     'temperature': 300,
     'debye_waller_grid': [6, 6, 6],
     'negative_e': True,
@@ -257,3 +308,18 @@ def test_sf_unit_change(material, opt_dict):
     n_atoms = w.shape[1]/6 # n_atoms is n_modes/6 because of 'negative_e'
     unit_scale = 1e11/(2*n_atoms)
     npt.assert_allclose(sf, expected_sf*unit_scale, rtol=1e-2, atol=1e-2)
+
+def test_invalid_fcs_raises_value_error():
+    with pytest.raises(ValueError):
+        euphonic_sqw_models.CoherentCrystal(None)
+
+@pytest.mark.parametrize('kwarg', [{'debye_waller': np.random.rand(5, 3, 3)},
+                                   {'debye_waller_grid': [2, 3]},
+                                   {'conversion_mat': np.random.rand(2, 3)},
+                                   {'temperature': -5*ureg('K')}])
+def test_invalid_kwargs_raises_value_error(kwarg):
+    fc = euphonic.ForceConstants.from_castep(
+        get_abspath('quartz.castep_bin', 'input'))
+    with pytest.raises(ValueError):
+        euphonic_sqw_models.CoherentCrystal(fc, **kwarg)
+